+++ packages/transformer-directives/src/functions.ts
@@ -1,6 +1,6 @@
 import type { FunctionNode, StringNode } from 'css-tree'
 import { transformThemeString } from '@unocss/rule-utils'
-import type { TransformerDirectivesContext } from '.'
+import type { TransformerDirectivesContext } from './types'
 
 export function handleFunction({ code, uno, options }: TransformerDirectivesContext, node: FunctionNode) {
   const { throwOnMissing = true } = options
